add start time and end time to api request
